requirements.txt
ffmpeg-python
librosa
numpy
soundfile
webdataset
rapidfuzz
megatron-lm
nemo_toolkit[all]
uvicorn[standard]
fastapi
python-multipart
